+++ src/results/extension-actions.ts
@@ -20,6 +20,7 @@ enum Commands {
   AskToDeleteJob = 'ask-to-delete-job',
   InitResults = 'init-results',
   UploadConf = 'upload-conf',
+  Rename = 'rename',
 }
 
 export function uploadConf(): void {
@@ -54,15 +55,15 @@ export function removeScript(name: string): void {
   })
 }
 
-export function stopScript(pid: number): void {
+export function stopScript(pid: number, modal: boolean): void {
   log({
     action: 'Send "stop-script" command',
     source: Sources.ResultsWebview,
-    info: pid,
+    info: { pid, modal },
   })
   vscode.postMessage({
     command: Commands.StopScript,
-    payload: pid,
+    payload: { pid: pid, modal: modal },
   })
 }
 
@@ -137,6 +138,21 @@ export function duplicate(
       toDuplicate: toDuplicate,
       duplicatedName: duplicated,
       rule: rule,
+    },
+  })
+}
+
+export function rename(oldName: JobNameMap, newName: JobNameMap): void {
+  log({
+    action: 'Send "rename" command',
+    source: Sources.ResultsWebview,
+    info: [oldName, newName],
+  })
+  vscode.postMessage({
+    command: Commands.Rename,
+    payload: {
+      oldName: oldName,
+      newName: newName,
     },
   })
 }