+++ src/settings/SpecSettingsTab.svelte
@@ -19,6 +19,7 @@
   } from './stores/store.js'
   import CheckBoxInfo from './components/CheckBoxInfo.svelte'
   import { manageFiles } from './utils/refreshFiles'
+  import { Source } from './types'
 
   // this items arrary contains all the solidity files and should update on when updateItems is fired
   // some fake stuff
@@ -222,6 +223,7 @@
                   placeholder="deafult: all rules"
                   bind:bindValue={$specObj.rules}
                   infoObj={infoObjArr.rules}
+                  source={Source.Spec}
                 />
               </div>
             </div>
@@ -241,6 +243,7 @@
                           placeholder="flag name"
                           bind:bindValue={obj.name}
                           infoObj={infoObjArr.flag}
+                          source={Source.Spec}
                         />
                       </div>
                       <div class="dark_input">
@@ -248,6 +251,7 @@
                           placeholder="value"
                           bind:bindValue={obj.value}
                           infoObj={infoObjArr.value}
+                          source={Source.Spec}
                         />
                       </div>
                       <i
@@ -267,9 +271,10 @@
                     <div class="dark_input input_x3">
                       <h3>Duration</h3>
                       <CustomInput
-                        placeholder="default: 600"
+                        placeholder="default: 300"
                         bind:bindValue={$specObj.duration}
                         infoObj={infoObjArr.duration}
+                        source={Source.Spec}
                       />
                     </div>
 
@@ -293,6 +298,7 @@
                         placeholder="default: 1"
                         bind:bindValue={$specObj.loopUnroll}
                         infoObj={infoObjArr.loop_iter}
+                        source={Source.Spec}
                       />
                     </div>
                   </div>
@@ -354,6 +360,20 @@
                     </div>
                   </div>
                   <div class="input_wrapper check_between  mt-8px">
+                    <div class="input_wrapper input_single">
+                      <div class="dark_input">
+                        <h3>Branch</h3>
+                        <CustomInput
+                          placeholder="default: master"
+                          bind:bindValue={$specObj.branchName}
+                          infoObj={infoObjArr.stg}
+                          source={Source.Spec}
+                        />
+                      </div>
+                    </div>
+                  </div>
+
+                  <div class="input_wrapper  mt-8px">
                     <div class="dark_input check_box_wrapper">
                       <label class="checkbox_container"
                         >Staging
@@ -366,18 +386,6 @@
                       <CheckBoxInfo infoObj={infoObjArr.stg} />
                     </div>
                   </div>
-                  {#if $specObj.runOnStg}
-                    <div class="input_wrapper input_single">
-                      <div class="dark_input">
-                        <h3>Branch Name</h3>
-                        <CustomInput
-                          placeholder="default: master"
-                          bind:bindValue={$specObj.branchName}
-                          infoObj={infoObjArr.stg}
-                        />
-                      </div>
-                    </div>
-                  {/if}
                 </div>
               </CollapseCard>
             </div>